+++ src/Security/UserChecker.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Security;
+
+use App\Entity\User;
+use Symfony\Component\Security\Core\Exception\CustomUserMessageAccountStatusException;
+use Symfony\Component\Security\Core\User\UserCheckerInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
+
+class UserChecker implements UserCheckerInterface
+{
+    public function checkPreAuth(UserInterface $user): void
+    {
+        if (!$user instanceof User) {
+            return;
+        }
+
+        if ($user->isDeleted()) {
+            throw new CustomUserMessageAccountStatusException('Ce compte a été supprimé.');
+        }
+    }
+
+    public function checkPostAuth(UserInterface $user)
+    {
+        // pas besoin de vérifier après connexion, on le fait déjà avant (checkPreAuth)
+    }
+
+}